requirements.txt
html2text
kivy.deps.sdl2
kivy.deps.glew
